Grouped issue worklog related tests. Tested "getWorklogs" method.
